pyproject.toml
[project]
name = "notion-automations"
version = "0.1.0"
description = "Automation scripts to enhance personal notion usage"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{name = "Namit Deb", email = "namitdeb739@gmail.com"}]
dependencies = [
    "ics>=0.7.2",
    "notion-client>=3.0.0",
    "questionary>=2.1.1",
    "typer>=0.12",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "jupyter>=1.0",
    "ruff>=0.5",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.25",
    "pip-audit>=2.7",
    "commitizen>=4.1",
    "copier>=9.0",
    "python-dotenv>=1.2.2",
]


[project.scripts]
notion-automations = "notion_automations.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/notion_automations"]

[tool.ruff]
target-version = "py311"
src = ["src"]
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "TC",  # flake8-type-checking
    "RUF", # ruff-specific rules
]

[tool.ruff.lint.isort]
known-first-party = ["notion_automations"]


[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"


[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.bumpversion]
current_version = "0.1.0"
commit = true
tag = true
message = "chore(release): bump version {current_version} → {new_version}"

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[[tool.bumpversion.files]]
filename = "src/notion_automations/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "v$version"
